% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gchartcolour_palettes.R
\name{assign_all_palettes}
\alias{assign_all_palettes}
\title{A function to assign all palettes}
\usage{
assign_all_palettes(palettes = NULL)
}
\arguments{
\item{palettes}{a list of palettes gained through get_g_chartcolour_palettes. If null, the palettes will be scraped. Defaults to NULL}
}
\description{
A function to assign all palettes
}
\examples{
assign_all_palettes()
}
