% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gchartcolour_palettes.R
\name{get_colours_from_palette}
\alias{get_colours_from_palette}
\title{A function to extract the swebsafe hex codes from a tabularised palette}
\usage{
get_colours_from_palette(palette)
}
\arguments{
\item{palette}{a formatted palette node in dataframe format, as returned by the above functions}
}
\description{
A function to extract the swebsafe hex codes from a tabularised palette
}
