% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gchartcolour_palettes.R
\name{get_g_chartcolour_palettes}
\alias{get_g_chartcolour_palettes}
\title{A function to scrape all the FT colour palettes. Returns a named list}
\usage{
get_g_chartcolour_palettes()
}
\description{
Returns a named list of the palettes on this page https://ft-interactive.github.io/g-chartcolour/
scrapes the palette nodes, then calls the get_palette function on each to return tabular results
}
\examples{
get_g_chartcolour_palettes()
}
