% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gchartcolour_palettes.R
\name{get_palette}
\alias{get_palette}
\title{A function to tabularise the data in a palette node.}
\usage{
get_palette(palette)
}
\arguments{
\item{palette}{a palette node from https://ft-interactive.github.io/g-chartcolour/}
}
\description{
This function tabularises a single palette. Is called only inside get_g_chartcolour_palettes
}
\examples{
get_palette()
}
