% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility_functions.R
\name{hex_three_to_hex_six}
\alias{hex_three_to_hex_six}
\title{Converts a three character hexcode, eg. #FFF to its six-character equivilent}
\usage{
hex_three_to_hex_six(hex)
}
\arguments{
\item{hex}{a websafe hexcode, starting with #}
}
\description{
This function is a wrapper around the paste(c, collapse = '')
}
