% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility_functions.R
\name{myCollapse}
\alias{myCollapse}
\title{A function to collapse a vector into a character string.}
\usage{
myCollapse(array, sep = "|")
}
\arguments{
\item{array}{a vector}

\item{sep}{what to interpace the array elements with, defualts to the or operator |}
}
\description{
This function is a wrapper around the paste(c, collapse = '')
}
