% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_g_chartcolour_palettes}
\alias{plot_g_chartcolour_palettes}
\title{A function to plot all the palettes in gchartcolour for colour picking}
\usage{
plot_g_chartcolour_palettes(g_chartcolour_palettes)
}
\arguments{
\item{g_chartcolour_palettes}{loaded palettes, using get_g_chartcolour_palettes}
}
\description{
This function plots all palettes using ggplot2. Requires the colour palettes to be in the global environment
}
