% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gchartcolour_palettes.R
\name{remove_all_palettes}
\alias{remove_all_palettes}
\title{A function to remove all palettes}
\usage{
remove_all_palettes(palettes)
}
\arguments{
\item{palettes}{a list of palettes gained through get_g_chartcolour_palettes.}
}
\description{
A function to remove all palettes
}
\examples{
remove_all_palettes()
}
